improvement: cut container size down
Currently the entire app and node_modules is shipped in the docker container,
however only the compiled version in /usr/src/app/dist is actually served.
This PR has a separate package.json and yarn.lock that is used so that
only the node modules required to run the inject.js and sirv are brought in.

As a side benefit to splitting the build process into two steps, it is easier
 to provide docker containers in other architectures, as only the steps
 to build the main container is duplicated, not the entire build process.

887MB vs 55.7MB according to `docker image history`